pyproject.toml
[tool.poetry]
name = "homework-webpython-12"
version = "0.1.0"
description = ""
authors = ["Андрій <andrey.maybrodskyy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
sqlalchemy = "^2.0.7"
alembic = "^1.10.2"
pydantic = {extras = ["email"], version = "^1.10.7"}
fastapi-jwt-auth = "^0.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
